+++ pkg/mtmd/chunks_test.go
@@ -61,7 +61,7 @@ func TestInputChunksGetType(t *testing.T) {
 		t.Fatalf("invalid chunk size: %d", size)
 	}
 
-	idx := uint32(1)
+	idx := uint64(1)
 	chunk := InputChunksGet(chunks, idx)
 	if chunk == InputChunk(0) {
 		t.Fatalf("InputChunksGet returned an invalid chunk for index %d", idx)
@@ -104,7 +104,7 @@ func TestInputChunkGetNTokens(t *testing.T) {
 
 	testSetupChunks(t, ctx, chunks)
 
-	idx := uint32(1)
+	idx := uint64(1)
 	chunk := InputChunksGet(chunks, idx)
 	if chunk == InputChunk(0) {
 		t.Fatalf("InputChunksGet returned an invalid chunk for index %d", idx)
@@ -141,7 +141,7 @@ func TestInputChunkGetId(t *testing.T) {
 
 	testSetupChunks(t, ctx, chunks)
 
-	idx := uint32(1)
+	idx := uint64(1)
 	chunk := InputChunksGet(chunks, idx)
 	if chunk == InputChunk(0) {
 		t.Fatalf("InputChunksGet returned an invalid chunk for index %d", idx)
@@ -178,7 +178,7 @@ func TestInputChunkGetNPos(t *testing.T) {
 
 	testSetupChunks(t, ctx, chunks)
 
-	idx := uint32(1)
+	idx := uint64(1)
 	chunk := InputChunksGet(chunks, idx)
 	if chunk == InputChunk(0) {
 		t.Fatalf("InputChunksGet returned an invalid chunk for index %d", idx)
@@ -215,7 +215,7 @@ func TestInputChunkCopyAndFree(t *testing.T) {
 
 	testSetupChunks(t, ctx, chunks)
 
-	idx := uint32(1)
+	idx := uint64(1)
 	chunk := InputChunksGet(chunks, idx)
 	if chunk == InputChunk(0) {
 		t.Fatalf("InputChunksGet returned an invalid chunk for index %d", idx)
@@ -259,7 +259,7 @@ func TestInputChunkGetTokensImage(t *testing.T) {
 
 	testSetupChunks(t, ctx, chunks)
 
-	idx := uint32(1)
+	idx := uint64(1)
 	chunk := InputChunksGet(chunks, idx)
 	if chunk == InputChunk(0) {
 		t.Fatalf("InputChunksGet returned an invalid chunk for index %d", idx)
